Controllers added: AttendanceController PayperiodController Models Added: Attendance_model Payperiod_model Some views added and trivial to specify.
On the main menu: added option: "Absences and Tardiness".

For generating the payroll, you should first generate absences and tardiness data, by accessing Home > Absences and Tardiness.

No UI still for result of the aforementioned, but you can see
it in PHPMYadmin.

There are a lot of new tables and revision to a few tables so
please import the SQL file first.